docs: Adds README.
This is an initial draft of what Banner is.  It describes influences
and outlines my initial stab at a DSL for CI/CD.

While the goals are unlikely to change, the DSL is likely to undergo
much iteration before it becomes a valid, usable language that covers
all of the edge cases that it will be required to address.