docs: Modifies a comment.
I was confusing what the engine needed to do with what the server
needs to do. The engines `execute` will take information about a
task and execute it within it's context.  I had started thinking
of it as this thing that understands pipelines and has a database
and needed to get stuff out and potentially do lots of stuff. This
should not be the case.