feat(parser): Adds imports
Adds imports to the grammar so that we can compose a pipeline
of many files; regardless of location.  I initially started
trying to add a propper URI grammar based on the grammar outlined
in RFC-3986 (https://www.rfc-editor.org/rfc/rfc3986#appendix-A)
and this would have been good, but it occured to me that at some
point I might want to use variables and interpolation in these
URI's so I ditched the idea, and just accept any old thing, and
later will parse the URI; after substitutions; with the iref crate.