feat(event-handlers): Adds event handling capability.
This started as what I thought would be a small, and relatively simple
fix. In the end, it was quite complicated, and required quite a bit of
refactoring on top of the new functionality. Happy with the result, as
you can now enact an entire disconnected pipeline; by that I mean, running
a series of serial tasks created by pipeline and job directives that
execute in sequence and trigger subsequent tasks, or of course stop if
there is a failure.

* Adds capability to run event handlers for pipeline and job directives.
* Creates event handlers that run with the Rune scripting engine.
* Fixed and added tests.