feat: Adds support for directory state
There is currently  a known issue with the COMMENTs that forces
pipeline parsing failures. An issue has been raised with the
pest peeps here: https://github.com/pest-parser/pest/issues/857

Banner can now conceivably dog food itself. Allowing directories
to be declared as engine supplied, and also being able to reference
them from other tasks, allows rudimentry state to be passed around
a pipeline. The bannar-pipeline.ban in pipeline assets has been
run and shown to clone the repo to a local directory in the pre-defined
state directory, and then the output used in the test task to build
and run the tests; 2 of which fail.. :( bad commit...

If you want to use this commit, currently, you'll need to change the
COMMENT to three slashes `///` at present this is just `//`. Without
the change, parsing will fail and the pipeline will not run.