fix(bannar-parser): Fixes COMMENT parsing.
There was a problem parsing what kind of looked like comments
inside of a string literal.  This was caused by the string literal
not being atomic.  Tried to change this and fix the ordering problem
that job and pipeline have in the parser, but fell short.  Will
fix these in a later commit.

Some tests are intentionally left broken.