feat(parser): Adds parallel and sequential tasks for jobs and pipelines
While the capability to parse a both is now available, much of the plumbing
work that is required to enable the execution of them has been left in a
"todo!()" state. I bounced around a lot between completely removing events
and event handlers and adding them back in. It would be significantly easier
to just generate a rune script with all the async awaits and joins in the
right place and let the script take care of it, but a great deal of funcitonality
would be lost if that were to happen.  So the events stay.

Next up will be figuring out how to hold state across multiple events. This
is needed in order to successfully complete a block of parallel tasks,
triggering the next sequential one.