feature(ui): Adds basic pipeline render to cli.
There's a load more to do here, but just being able to visualise the
pipeline, however basic, is a great thing to have. The rendering
doesn't update yet based on the current status of a job, and this
should be one of the next things to be done. However the whole render
is impared by the lack of parallel execution for jobs; this should be
the same as for tasks but has not been implemented yet.

* Added sequential pipeline rendering.
* Fiddled with the full screen rendering of logs and events, removing
  borders to allow selection of text without border gumpf.
* Updated Ratatui.

Next steps:
* Fix pipeline execution of parallel jobs.
* Ensure parallel jobs render correctly.
* Copy pipeline/job structures to the engine and add status.
* Wrap engine pipeline/job structs in one that works with the UI
  to record postions.