fix(engine): Fixes the job triggering.
I'm not sure if this is going to be everything, but the test pipeline
is now pretty complex, and all the jobs; simple tho they may be; trigger
in the right order, and render out just fine. I'll do more rigourous testing
of the rendering AND the job triggering soon.