feat(ui): Colourises job connectors
Connectors from previous jobs now show in the colour of the job being connected
to; right side. I personally find this more effective a means of directing my
attention to running and failed jobs.

Made the Banner pipeline more of a one job per task implementation so as to
capture the flow better. This will possibly be reverted at some point in the
future when the pipeline becomes more complex.