feat(parser, engine): Defines job macros
The upshot of this feature is that you can now specify a job name in
a pipeline definition as a macro (of sorts). This means you no longer
need to additionally specify the job if it only ever calls straight
through to a task of the same name.

The syntax for this is an exclaimation mark `!` after the job name in
in the pipeline. Like so:
```
pipeline my-pipeline [
  get-source!,
  unit-test,
  ...
]
```

where in this case, the `get-source` job has no corresponding job definition
that would have looked like this:
```
job get-source [
  get-source
]
```

Validation that this isn't then overwriting an existing job definition is yet
to come.