chore(engine): Consolidates job_success and job_fail events
This is just step one in a larger plan to reduce the complexity inherent in
the sheer number of event handlers for even moderate sized pipelines.
Now where there would have been two event handlers for a job completion;
one for success, and one for fail; there will now be only one.  Not only
is it a reduction in the number of handlers, but it is also more correct.
Previously only success and failure were catered for.  Now all Task completion
types are dealt with; including aborted and errored.

The ultimate aim is to reduce the number of event handlers for jobs to one for
the entire pipeline, and the same for Tasks and pipelines. This will of course
come at the cost of some additional memory or CPU. I'm not sure if this is even
possible yet, as I've not given a great deal of thought to how this interplays
with UserDefined event handlers.