feat(parser): Adds pragmas.
Pragmas can be defined (global scope only) in a pipeline file. This allows
for the engine implementations, to take a list of pragmas with specific
contexts and parse them as they see fit.

The need for this arose from needing a way to host expected outcomes with
the pipelines for the test-engine. Alternatives that were considered were:
* additional file (test.ban & test.outcomes); I didn't like this so much
  as there's some additional cognitive overhead. It might have been relatively
  easy to implement, but then then any user writing tests would have to switch
  between pipeline file and outcomes file in order to build a test.
* just define the outcomes in a rust test... I didn't like this as it meant
  pretty much the same context switching as an additional file, and meant that
  instead of just writing a loop over the files, a test would need to be created
  for each file.

Ultimately the pragmas felt like they had additional utility beyond just testing
and provided a clean enough mechanism for adding in a bunch of asserts on the
events produced by a pipeline run in a test. Not that I've reached there yet, but
I can mentally extend pragmas into the engine space for doing things like providing
default resource configurations for k8s/docker or other global metadata that
wouldn't fit neatly into a tag.