feat(engine): Exposes Pragmas to Engines.
Follow on from the last commit that enabled parsing of pragmas in a pipeline.
Thi commit exposes the pragmas to the engines via a builder, which allows an
engine to register its interest in pragma contexts and thus receive all registered
contexts. It's then up to the engine to do what it wants with the pragmas it gets.

I think a future enhancement to this would be to add a parser implementation to
each context so that the pragmas can be parsed and validated before being handed
back to the engine. The engine can then guarantee that the pragmas is has are valid
as per the specification it has for them instead of having to post process the pragmas.