chore: Updates Cargo.toml files using workspace dependencies for common crates
The adopted philosophy here is if the dependency is used by more than one crate
it should have an entry in the workspace toml file. Anything that is used only
by a single crate can exist soley in it's workspace, but it can exist in the
root toml file if we believe it will be used by more than one workspace in
future.