feat(engine): Makes channels broadcast instead of mpsc.
The test engine needs to be able to listen in to the raised events that the
local engine that it wraps receives. Might have been able to do this differently
but I have been thinking for a while it would be handy to have the channel
as an multi-producer/multi-consumer type, so done now.  And it should all still
work as expected.

Turns out that the blocking send on the events wasn't being used, so this got to
go; less code. \o/

Also had to remove the rustfmt setting that is nightly only. :( This could/is
preventing the banner pipeline itself from completing correctly.