chore: Updates Ratatui.
Updates the Ratatui crate, and as a result needed to use the git repo for the logger
I'll unwind the logger change when they publish the crate.

Also pinned the Rune crate, as newer changes break some of the work that has been done.
I need to control the update when it happens and limit it to just the changes for
that update, which I feel will be relatively extensive to the code it affects.