fix(ui): Fixes the auto reset of the layout to multi-panel
When I did the clippy refactorings to fix up all the pedantic lints, I moved
the ui state struct inside the main ui loop. This had the effect of resetting
the layout everytime you tried to go full screen on one of the panels. I was pretty
swamped at the time with lints, and didn't think much of the change. All the tests
passed, and when I ran the cli, I only looked at the multi-panel default view.
So I introduced a logic error. :(

Turns out the actual fix was pretty simple, just derive `Copy` for UiState. Which
is perfectly fine in this situation, because the state is only ever mutated in the
main loop before being passed into the rendering functions that only ever read the
state.